split (L)RM(S) out of agent, get fork working
