+++ src/radical/pilot/agent/__init__.py
@@ -1,0 +1,3 @@
+
+from .rm import RM
+
