+++ src/radical/pilot/agent/radical-pilot-agent-multicore.py
@@ -218,17 +218,6 @@ LAUNCH_METHOD_RUNJOB        = 'RUNJOB'
 LAUNCH_METHOD_SSH           = 'SSH'
 LAUNCH_METHOD_YARN          = 'YARN'
 
-# 'enum' for local resource manager types
-LRMS_NAME_CCM               = 'CCM'
-LRMS_NAME_FORK              = 'FORK'
-LRMS_NAME_LOADLEVELER       = 'LOADL'
-LRMS_NAME_LSF               = 'LSF'
-LRMS_NAME_PBSPRO            = 'PBSPRO'
-LRMS_NAME_SGE               = 'SGE'
-LRMS_NAME_SLURM             = 'SLURM'
-LRMS_NAME_TORQUE            = 'TORQUE'
-LRMS_NAME_YARN              = 'YARN'
-
 # 'enum' for pilot's unit scheduler types
 SCHEDULER_NAME_CONTINUOUS   = "CONTINUOUS"
 SCHEDULER_NAME_SCATTERED    = "SCATTERED"
@@ -2788,7 +2777,7 @@ class LaunchMethodYARN(LaunchMethod):
         # the first node in the list and the port is the default one, else
         # it is the one that the YARN LRMS returns
         hadoop_home = None
-        if lrms.name == 'YARNLRMS':
+        if lrms.name == 'YARNLRMS': # FIXME: use constant
             logger.info('Hook called by YARN LRMS')
             logger.info('NameNode: {0}'.format(lrms.namenode_url))
             service_url    = lrms.namenode_url
@@ -2900,7 +2889,7 @@ class LaunchMethodYARN(LaunchMethod):
         if 'name' not in lm_info:
             raise RuntimeError('name not in lm_info for %s' % name)
 
-        if lm_info['name'] != 'YARNLRMS':
+        if lm_info['name'] != 'YARNLRMS': # FIXME: use constant
             logger.info('Stoping YARN')
             os.system(lm_info['hadoop_home'] + '/sbin/stop-yarn.sh')
 
@@ -3046,1478 +3035,6 @@ class LaunchMethodYARN(LaunchMethod):
         self._log.debug("Yarn Command %s"%yarn_command)
 
         return print_str+yarn_command, None
-
-
-
-# ==============================================================================
-#
-# Base class for LRMS implementations.
-#
-# ==============================================================================
-#
-class LRMS(object):
-    """
-    The Local Resource Manager (LRMS -- where does the 's' come from, actually?)
-    provide three fundamental information:
-
-      LRMS.node_list      : a list of node names
-      LRMS.agent_node_list: the list of nodes reserved for agent execution
-      LRMS.cores_per_node : the number of cores each node has available
-
-    Schedulers can rely on these information to be available.  Specific LRMS
-    incarnation may have additional information available -- but schedulers
-    relying on those are invariably bound to the specific LRMS.  An example is
-    the Torus Scheduler which relies on detailed torus layout information from
-    the LoadLevelerLRMS (which describes the BG/Q).
-
-    The LRMS will reserve nodes for the agent execution, by deriving the
-    respectively required node count from the config's agent_layout section.
-    Those nodes will be listed in LRMS.agent_node_list. Schedulers MUST NOT use
-    the agent_node_list to place compute units -- CUs are limited to the nodes
-    in LRMS.node_list.
-
-    Additionally, the LRMS can inform the agent about the current hostname
-    (LRMS.hostname()) and ip (LRMS.hostip()).  Once we start to spread the agent
-    over some compute nodes, we may want to block the respective nodes on LRMS
-    level, so that is only reports the remaining nodes to the scheduler.
-    """
-
-    # TODO: Core counts dont have to be the same number for all hosts.
-
-    # TODO: We might not have reserved the whole node.
-
-    # TODO: Given that the Agent can determine the real core count, in
-    #       principle we could just ignore the config and use as many as we
-    #       have to our availability (taken into account that we might not
-    #       have the full node reserved of course)
-    #       Answer: at least on Yellowstone this doesnt work for MPI,
-    #               as you can't spawn more tasks then the number of slots.
-
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        self.name            = type(self).__name__
-        self._cfg            = cfg
-        self._log            = logger
-        self.requested_cores = self._cfg['cores']
-
-        self._log.info("Configuring LRMS %s.", self.name)
-
-        self.lm_info         = dict()
-        self.lrms_info       = dict()
-        self.slot_list       = list()
-        self.node_list       = list()
-        self.agent_nodes     = {}
-        self.cores_per_node  = None
-
-        # The LRMS will possibly need to reserve nodes for the agent, according to the
-        # agent layout.  We dig out the respective requirements from the config
-        # right here.
-        self._agent_reqs = []
-        layout = self._cfg['agent_layout']
-        # FIXME: this loop iterates over all agents *defined* in the layout, not
-        #        over all agents which are to be actually executed, thus
-        #        potentially reserving too many nodes.
-        for worker in layout:
-            target = layout[worker].get('target')
-            # make sure that the target either 'local', which we will ignore,
-            # or 'node'.
-            if target == 'local':
-                pass # ignore that one
-            elif target == 'node':
-                self._agent_reqs.append(worker)
-            else :
-                raise ValueError("ill-formatted agent target '%s'" % target)
-
-        # We are good to get rolling, and to detect the runtime environment of
-        # the local LRMS.
-        self._configure()
-        logger.info("Discovered execution environment: %s", self.node_list)
-
-        # Make sure we got a valid nodelist and a valid setting for
-        # cores_per_node
-        if not self.node_list or self.cores_per_node < 1:
-            raise RuntimeError('LRMS configuration invalid (%s)(%s)' % \
-                    (self.node_list, self.cores_per_node))
-
-        # Check if the LRMS implementation reserved agent nodes.  If not, pick
-        # the first couple of nodes from the nodelist as a fallback.
-        if self._agent_reqs and not self.agent_nodes:
-            self._log.info('Determine list of agent nodes generically.')
-            for worker in self._agent_reqs:
-                # Get a node from the end of the node list
-                self.agent_nodes[worker] = self.node_list.pop()
-                # If all nodes are taken by workers now, we can safely stop,
-                # and let the raise below do its thing.
-                if not self.node_list:
-                    break
-
-        if self.agent_nodes:
-            self._log.info('Reserved agent node(s): %s' % self.agent_nodes.values())
-            self._log.info('Agent(s) running on node(s): %s' % self.agent_nodes.keys())
-            self._log.info('Remaining work node(s): %s' % self.node_list)
-
-        # Check if we can do any work
-        if not self.node_list:
-            raise RuntimeError('LRMS has no nodes left to run units')
-
-        # After LRMS configuration, we call any existing config hooks on the
-        # launch methods.  Those hooks may need to adjust the LRMS settings
-        # (hello ORTE).  We only call LM hooks *once*
-        launch_methods = set() # set keeps entries unique
-        if 'mpi_launch_method' in self._cfg:
-            launch_methods.add(self._cfg['mpi_launch_method'])
-        launch_methods.add(self._cfg['task_launch_method'])
-        launch_methods.add(self._cfg['agent_launch_method'])
-
-        for lm in launch_methods:
-            if lm:
-                try:
-                    ru.dict_merge(self.lm_info,
-                            LaunchMethod.lrms_config_hook(lm, self._cfg, self, self._log))
-                except Exception as e:
-                    self._log.exception("lrms config hook failed")
-                    raise
-
-                self._log.info("lrms config hook succeeded (%s)" % lm)
-
-        # For now assume that all nodes have equal amount of cores
-        cores_avail = (len(self.node_list) + len(self.agent_nodes)) * self.cores_per_node
-        if 'RADICAL_PILOT_PROFILE' not in os.environ:
-            if cores_avail < int(self.requested_cores):
-                raise ValueError("Not enough cores available (%s) to satisfy allocation request (%s)." \
-                                % (str(cores_avail), str(self.requested_cores)))
-
-        # NOTE: self.lrms_info is what scheduler and launch method can
-        # ultimately use, as it is included into the cfg passed to all
-        # components.
-        #
-        # four elements are well defined:
-        #   lm_info:        the dict received via the LM's lrms_config_hook
-        #   node_list:      a list of node names to be used for unit execution
-        #   cores_per_node: as the name says
-        #   agent_nodes:    list of node names reserved for agent execution
-        #
-        # That list may turn out to be insufficient for some schedulers.  Yarn
-        # for example may need to communicate YARN service endpoints etc.  an
-        # LRMS can thus expand this dict, but is then likely bound to a specific
-        # scheduler which can interpret the additional information.
-        self.lrms_info['lm_info']        = self.lm_info
-        self.lrms_info['node_list']      = self.node_list
-        self.lrms_info['cores_per_node'] = self.cores_per_node
-        self.lrms_info['agent_nodes']    = self.agent_nodes
-
-
-    # --------------------------------------------------------------------------
-    #
-    # This class-method creates the appropriate sub-class for the LRMS.
-    #
-    @classmethod
-    def create(cls, name, cfg, logger):
-
-        # Make sure that we are the base-class!
-        if cls != LRMS:
-            raise TypeError("LRMS Factory only available to base class!")
-
-        try:
-            impl = {
-                LRMS_NAME_CCM         : CCMLRMS,
-                LRMS_NAME_FORK        : ForkLRMS,
-                LRMS_NAME_LOADLEVELER : LoadLevelerLRMS,
-                LRMS_NAME_LSF         : LSFLRMS,
-                LRMS_NAME_PBSPRO      : PBSProLRMS,
-                LRMS_NAME_SGE         : SGELRMS,
-                LRMS_NAME_SLURM       : SLURMLRMS,
-                LRMS_NAME_TORQUE      : TORQUELRMS,
-                LRMS_NAME_YARN        : YARNLRMS
-            }[name]
-            return impl(cfg, logger)
-
-        except KeyError:
-            logger.exception('lrms construction error')
-            raise RuntimeError("LRMS type '%s' unknown or defunct" % name)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def stop(self):
-
-        # During LRMS termination, we call any existing shutdown hooks on the
-        # launch methods.  We only call LM shutdown hooks *once*
-        launch_methods = set() # set keeps entries unique
-        if 'mpi_launch_method' in self._cfg:
-            launch_methods.add(self._cfg['mpi_launch_method'])
-        launch_methods.add(self._cfg['task_launch_method'])
-        launch_methods.add(self._cfg['agent_launch_method'])
-
-        for lm in launch_methods:
-            if lm:
-                try:
-                    LaunchMethod.lrms_shutdown_hook(lm, self._cfg, self,
-                                                    self.lm_info, self._log)
-                except Exception as e:
-                    self._log.exception("lrms shutdown hook failed")
-                    raise
-
-                self._log.info("lrms shutdown hook succeeded (%s)" % lm)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        raise NotImplementedError("_Configure not implemented for LRMS type: %s." % self.name)
-
-
-    # --------------------------------------------------------------------------
-    #
-    @staticmethod
-    def hostip(req=None, logger=None):
-        """
-        Look up the ip number for a given requested interface name.
-        If interface is not given, do some magic.
-        """
-
-        # List of interfaces that we probably dont want to bind to by default
-        black_list = ['lo', 'sit0']
-
-        # Known intefaces in preferred order
-        sorted_preferred = [
-            'ipogif0', # Cray's
-            'br0', # SuperMIC
-            'eth0'
-        ]
-
-        # Get a list of all network interfaces
-        all = netifaces.interfaces()
-
-        logger.debug("Network interfaces detected: %s", all)
-
-        pref = None
-        # If we got a request, see if it is in the list that we detected
-        if req and req in all:
-            # Requested is available, set it
-            pref = req
-        else:
-            # No requested or request not found, create preference list
-            potentials = [iface for iface in all if iface not in black_list]
-
-        # If we didn't select an interface already
-        if not pref:
-            # Go through the sorted list and see if it is available
-            for iface in sorted_preferred:
-                if iface in all:
-                    # Found something, get out of here
-                    pref = iface
-                    break
-
-        # If we still didn't find something, grab the first one from the
-        # potentials if it has entries
-        if not pref and potentials:
-            pref = potentials[0]
-
-        # If there were no potentials, see if we can find one in the blacklist
-        if not pref:
-            for iface in black_list:
-                if iface in all:
-                    pref = iface
-
-        # Use IPv4, because, we can ...
-        af = netifaces.AF_INET
-        ip = netifaces.ifaddresses(pref)[af][0]['addr']
-
-        return ip
-
-
-
-# ==============================================================================
-#
-class CCMLRMS(LRMS):
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LRMS.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        self._log.info("Configured to run on system with %s.", self.name)
-
-        CCM_NODEFILE_DIR = os.path.expanduser('~/.crayccm')
-
-        ccm_nodefile_list = filter(lambda x: x.startswith('ccm_nodelist'),
-                                   os.listdir(CCM_NODEFILE_DIR))
-        if not ccm_nodefile_list:
-            raise Exception("No CCM nodefiles found in: %s." % CCM_NODEFILE_DIR)
-
-        ccm_nodefile_name = max(ccm_nodefile_list, key=lambda x:
-                              os.stat(os.path.join(CCM_NODEFILE_DIR, x)).st_mtime)
-        ccm_nodefile = os.path.join(CCM_NODEFILE_DIR, ccm_nodefile_name)
-
-        hostname = os.uname()[1]
-        if not hostname in open(ccm_nodefile).read():
-            raise RuntimeError("Using the most recent CCM nodefile (%s),"
-                               " but I (%s) am not in it!" % (ccm_nodefile, hostname))
-
-        # Parse the CCM nodefile
-        ccm_nodes = [line.strip() for line in open(ccm_nodefile)]
-        self._log.info("Found CCM nodefile: %s.", ccm_nodefile)
-
-        # Get the number of raw entries
-        ccm_nodes_length = len(ccm_nodes)
-
-        # Unique nodes
-        ccm_node_list = list(set(ccm_nodes))
-        ccm_node_list_length = len(ccm_node_list)
-
-        # Some simple arithmetic
-        self.cores_per_node = ccm_nodes_length / ccm_node_list_length
-
-        self.node_list = ccm_node_list
-
-
-# ==============================================================================
-#
-class TORQUELRMS(LRMS):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LRMS.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        self._log.info("Configured to run on system with %s.", self.name)
-
-        torque_nodefile = os.environ.get('PBS_NODEFILE')
-        if torque_nodefile is None:
-            msg = "$PBS_NODEFILE not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        # Parse PBS the nodefile
-        torque_nodes = [line.strip() for line in open(torque_nodefile)]
-        self._log.info("Found Torque PBS_NODEFILE %s: %s", torque_nodefile, torque_nodes)
-
-        # Number of cpus involved in allocation
-        val = os.environ.get('PBS_NCPUS')
-        if val:
-            torque_num_cpus = int(val)
-        else:
-            msg = "$PBS_NCPUS not set! (new Torque version?)"
-            torque_num_cpus = None
-            self._log.warning(msg)
-
-        # Number of nodes involved in allocation
-        val = os.environ.get('PBS_NUM_NODES')
-        if val:
-            torque_num_nodes = int(val)
-        else:
-            msg = "$PBS_NUM_NODES not set! (old Torque version?)"
-            torque_num_nodes = None
-            self._log.warning(msg)
-
-        # Number of cores (processors) per node
-        val = os.environ.get('PBS_NUM_PPN')
-        if val:
-            torque_cores_per_node = int(val)
-        else:
-            msg = "$PBS_NUM_PPN is not set!"
-            torque_cores_per_node = None
-            self._log.warning(msg)
-
-        if torque_cores_per_node in [None, 1]:
-            # lets see if SAGA has been forthcoming with some information
-            self._log.warning("fall back to $SAGA_PPN : %s", os.environ.get ('SAGA_PPN', None))
-            torque_cores_per_node = int(os.environ.get('SAGA_PPN', torque_cores_per_node))
-
-        # Number of entries in nodefile should be PBS_NUM_NODES * PBS_NUM_PPN
-        torque_nodes_length = len(torque_nodes)
-        torque_node_list = []
-        [torque_node_list.append(i) for i in torque_nodes if not torque_node_list.count(i)]
-
-      # if torque_num_nodes and torque_cores_per_node and \
-      #     torque_nodes_length < torque_num_nodes * torque_cores_per_node:
-      #     msg = "Number of entries in $PBS_NODEFILE (%s) does not match with $PBS_NUM_NODES*$PBS_NUM_PPN (%s*%s)" % \
-      #           (torque_nodes_length, torque_num_nodes,  torque_cores_per_node)
-      #     raise RuntimeError(msg)
-
-        # only unique node names
-        torque_node_list_length = len(torque_node_list)
-        self._log.debug("Node list: %s(%d)", torque_node_list, torque_node_list_length)
-
-        if torque_num_nodes and torque_cores_per_node:
-            # Modern style Torque
-            self.cores_per_node = torque_cores_per_node
-        elif torque_num_cpus:
-            # Blacklight style (TORQUE-2.3.13)
-            self.cores_per_node = torque_num_cpus
-        else:
-            # Old style Torque (Should we just use this for all versions?)
-            self.cores_per_node = torque_nodes_length / torque_node_list_length
-        self.node_list = torque_node_list
-
-
-# ==============================================================================
-#
-class PBSProLRMS(LRMS):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LRMS.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # TODO: $NCPUS?!?! = 1 on archer
-
-        pbspro_nodefile = os.environ.get('PBS_NODEFILE')
-
-        if pbspro_nodefile is None:
-            msg = "$PBS_NODEFILE not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        self._log.info("Found PBSPro $PBS_NODEFILE %s." % pbspro_nodefile)
-
-        # Dont need to parse the content of nodefile for PBSPRO, only the length
-        # is interesting, as there are only duplicate entries in it.
-        pbspro_nodes_length = len([line.strip() for line in open(pbspro_nodefile)])
-
-        # Number of Processors per Node
-        val = os.environ.get('NUM_PPN')
-        if val:
-            pbspro_num_ppn = int(val)
-        else:
-            msg = "$NUM_PPN not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        # Number of Nodes allocated
-        val = os.environ.get('NODE_COUNT')
-        if val:
-            pbspro_node_count = int(val)
-        else:
-            msg = "$NODE_COUNT not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        # Number of Parallel Environments
-        val = os.environ.get('NUM_PES')
-        if val:
-            pbspro_num_pes = int(val)
-        else:
-            msg = "$NUM_PES not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        pbspro_vnodes = self._parse_pbspro_vnodes()
-
-        # Verify that $NUM_PES == $NODE_COUNT * $NUM_PPN == len($PBS_NODEFILE)
-        if not (pbspro_node_count * pbspro_num_ppn == pbspro_num_pes == pbspro_nodes_length):
-            self._log.warning("NUM_PES != NODE_COUNT * NUM_PPN != len($PBS_NODEFILE)")
-
-        self.cores_per_node = pbspro_num_ppn
-        self.node_list = pbspro_vnodes
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _parse_pbspro_vnodes(self):
-
-        # PBS Job ID
-        val = os.environ.get('PBS_JOBID')
-        if val:
-            pbspro_jobid = val
-        else:
-            msg = "$PBS_JOBID not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        # Get the output of qstat -f for this job
-        output = subprocess.check_output(["qstat", "-f", pbspro_jobid])
-
-        # Get the (multiline) 'exec_vnode' entry
-        vnodes_str = ''
-        for line in output.splitlines():
-            # Detect start of entry
-            if 'exec_vnode = ' in line:
-                vnodes_str += line.strip()
-            elif vnodes_str:
-                # Find continuing lines
-                if " = " not in line:
-                    vnodes_str += line.strip()
-                else:
-                    break
-
-        # Get the RHS of the entry
-        rhs = vnodes_str.split('=',1)[1].strip()
-        self._log.debug("input: %s", rhs)
-
-        nodes_list = []
-        # Break up the individual node partitions into vnode slices
-        while True:
-            idx = rhs.find(')+(')
-
-            node_str = rhs[1:idx]
-            nodes_list.append(node_str)
-            rhs = rhs[idx+2:]
-
-            if idx < 0:
-                break
-
-        vnodes_list = []
-        cpus_list = []
-        # Split out the slices into vnode name and cpu count
-        for node_str in nodes_list:
-            slices = node_str.split('+')
-            for _slice in slices:
-                vnode, cpus = _slice.split(':')
-                cpus = int(cpus.split('=')[1])
-                self._log.debug("vnode: %s cpus: %s", vnode, cpus)
-                vnodes_list.append(vnode)
-                cpus_list.append(cpus)
-
-        self._log.debug("vnodes: %s", vnodes_list)
-        self._log.debug("cpus: %s", cpus_list)
-
-        cpus_list = list(set(cpus_list))
-        min_cpus = int(min(cpus_list))
-
-        if len(cpus_list) > 1:
-            self._log.debug("Detected vnodes of different sizes: %s, the minimal is: %d.", cpus_list, min_cpus)
-
-        node_list = []
-        for vnode in vnodes_list:
-            # strip the last _0 of the vnodes to get the node name
-            node_list.append(vnode.rsplit('_', 1)[0])
-
-        # only unique node names
-        node_list = list(set(node_list))
-        self._log.debug("Node list: %s", node_list)
-
-        # Return the list of node names
-        return node_list
-
-
-
-# ==============================================================================
-#
-class SLURMLRMS(LRMS):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LRMS.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        slurm_nodelist = os.environ.get('SLURM_NODELIST')
-        if slurm_nodelist is None:
-            msg = "$SLURM_NODELIST not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        # Parse SLURM nodefile environment variable
-        slurm_nodes = hostlist.expand_hostlist(slurm_nodelist)
-        self._log.info("Found SLURM_NODELIST %s. Expanded to: %s", slurm_nodelist, slurm_nodes)
-
-        # $SLURM_NPROCS = Total number of cores allocated for the current job
-        slurm_nprocs_str = os.environ.get('SLURM_NPROCS')
-        if slurm_nprocs_str is None:
-            msg = "$SLURM_NPROCS not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-        else:
-            slurm_nprocs = int(slurm_nprocs_str)
-
-        # $SLURM_NNODES = Total number of (partial) nodes in the job's resource allocation
-        slurm_nnodes_str = os.environ.get('SLURM_NNODES')
-        if slurm_nnodes_str is None:
-            msg = "$SLURM_NNODES not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-        else:
-            slurm_nnodes = int(slurm_nnodes_str)
-
-        # $SLURM_CPUS_ON_NODE = Number of cores per node (physically)
-        slurm_cpus_on_node_str = os.environ.get('SLURM_CPUS_ON_NODE')
-        if slurm_cpus_on_node_str is None:
-            msg = "$SLURM_CPUS_ON_NODE not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-        else:
-            slurm_cpus_on_node = int(slurm_cpus_on_node_str)
-
-        # Verify that $SLURM_NPROCS <= $SLURM_NNODES * $SLURM_CPUS_ON_NODE
-        if not slurm_nprocs <= slurm_nnodes * slurm_cpus_on_node:
-            self._log.warning("$SLURM_NPROCS(%d) <= $SLURM_NNODES(%d) * $SLURM_CPUS_ON_NODE(%d)",
-                            slurm_nprocs, slurm_nnodes, slurm_cpus_on_node)
-
-        # Verify that $SLURM_NNODES == len($SLURM_NODELIST)
-        if slurm_nnodes != len(slurm_nodes):
-            self._log.error("$SLURM_NNODES(%d) != len($SLURM_NODELIST)(%d)",
-                           slurm_nnodes, len(slurm_nodes))
-
-        # Report the physical number of cores or the total number of cores
-        # in case of a single partial node allocation.
-        self.cores_per_node = min(slurm_cpus_on_node, slurm_nprocs)
-
-        self.node_list = slurm_nodes
-
-
-
-# ==============================================================================
-#
-class SGELRMS(LRMS):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LRMS.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        sge_hostfile = os.environ.get('PE_HOSTFILE')
-        if sge_hostfile is None:
-            msg = "$PE_HOSTFILE not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        # SGE core configuration might be different than what multiprocessing
-        # announces
-        # Alternative: "qconf -sq all.q|awk '/^slots *[0-9]+$/{print $2}'"
-
-        # Parse SGE hostfile for nodes
-        sge_node_list = [line.split()[0] for line in open(sge_hostfile)]
-        # Keep only unique nodes
-        sge_nodes = list(set(sge_node_list))
-        self._log.info("Found PE_HOSTFILE %s. Expanded to: %s", sge_hostfile, sge_nodes)
-
-        # Parse SGE hostfile for cores
-        sge_cores_count_list = [int(line.split()[1]) for line in open(sge_hostfile)]
-        sge_core_counts = list(set(sge_cores_count_list))
-        sge_cores_per_node = min(sge_core_counts)
-        self._log.info("Found unique core counts: %s Using: %d", sge_core_counts, sge_cores_per_node)
-
-        self.node_list = sge_nodes
-        self.cores_per_node = sge_cores_per_node
-
-
-
-# ==============================================================================
-#
-class LSFLRMS(LRMS):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LRMS.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        lsf_hostfile = os.environ.get('LSB_DJOB_HOSTFILE')
-        if lsf_hostfile is None:
-            msg = "$LSB_DJOB_HOSTFILE not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        lsb_mcpu_hosts = os.environ.get('LSB_MCPU_HOSTS')
-        if lsb_mcpu_hosts is None:
-            msg = "$LSB_MCPU_HOSTS not set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        # parse LSF hostfile
-        # format:
-        # <hostnameX>
-        # <hostnameX>
-        # <hostnameY>
-        # <hostnameY>
-        #
-        # There are in total "-n" entries (number of tasks)
-        # and "-R" entries per host (tasks per host).
-        # (That results in "-n" / "-R" unique hosts)
-        #
-        lsf_nodes = [line.strip() for line in open(lsf_hostfile)]
-        self._log.info("Found LSB_DJOB_HOSTFILE %s. Expanded to: %s",
-                      lsf_hostfile, lsf_nodes)
-        lsf_node_list = list(set(lsf_nodes))
-
-        # Grab the core (slot) count from the environment
-        # Format: hostX N hostY N hostZ N
-        lsf_cores_count_list = map(int, lsb_mcpu_hosts.split()[1::2])
-        lsf_core_counts = list(set(lsf_cores_count_list))
-        lsf_cores_per_node = min(lsf_core_counts)
-        self._log.info("Found unique core counts: %s Using: %d",
-                      lsf_core_counts, lsf_cores_per_node)
-
-        self.node_list = lsf_node_list
-        self.cores_per_node = lsf_cores_per_node
-
-
-
-# ==============================================================================
-#
-class LoadLevelerLRMS(LRMS):
-
-    # --------------------------------------------------------------------------
-    #
-    # BG/Q Topology of Nodes within a Board
-    #
-    BGQ_BOARD_TOPO = {
-        0: {'A': 29, 'B':  3, 'C':  1, 'D': 12, 'E':  7},
-        1: {'A': 28, 'B':  2, 'C':  0, 'D': 13, 'E':  6},
-        2: {'A': 31, 'B':  1, 'C':  3, 'D': 14, 'E':  5},
-        3: {'A': 30, 'B':  0, 'C':  2, 'D': 15, 'E':  4},
-        4: {'A': 25, 'B':  7, 'C':  5, 'D':  8, 'E':  3},
-        5: {'A': 24, 'B':  6, 'C':  4, 'D':  9, 'E':  2},
-        6: {'A': 27, 'B':  5, 'C':  7, 'D': 10, 'E':  1},
-        7: {'A': 26, 'B':  4, 'C':  6, 'D': 11, 'E':  0},
-        8: {'A': 21, 'B': 11, 'C':  9, 'D':  4, 'E': 15},
-        9: {'A': 20, 'B': 10, 'C':  8, 'D':  5, 'E': 14},
-        10: {'A': 23, 'B':  9, 'C': 11, 'D':  6, 'E': 13},
-        11: {'A': 22, 'B':  8, 'C': 10, 'D':  7, 'E': 12},
-        12: {'A': 17, 'B': 15, 'C': 13, 'D':  0, 'E': 11},
-        13: {'A': 16, 'B': 14, 'C': 12, 'D':  1, 'E': 10},
-        14: {'A': 19, 'B': 13, 'C': 15, 'D':  2, 'E':  9},
-        15: {'A': 18, 'B': 12, 'C': 14, 'D':  3, 'E':  8},
-        16: {'A': 13, 'B': 19, 'C': 17, 'D': 28, 'E': 23},
-        17: {'A': 12, 'B': 18, 'C': 16, 'D': 29, 'E': 22},
-        18: {'A': 15, 'B': 17, 'C': 19, 'D': 30, 'E': 21},
-        19: {'A': 14, 'B': 16, 'C': 18, 'D': 31, 'E': 20},
-        20: {'A':  9, 'B': 23, 'C': 21, 'D': 24, 'E': 19},
-        21: {'A':  8, 'B': 22, 'C': 20, 'D': 25, 'E': 18},
-        22: {'A': 11, 'B': 21, 'C': 23, 'D': 26, 'E': 17},
-        23: {'A': 10, 'B': 20, 'C': 22, 'D': 27, 'E': 16},
-        24: {'A':  5, 'B': 27, 'C': 25, 'D': 20, 'E': 31},
-        25: {'A':  4, 'B': 26, 'C': 24, 'D': 21, 'E': 30},
-        26: {'A':  7, 'B': 25, 'C': 27, 'D': 22, 'E': 29},
-        27: {'A':  6, 'B': 24, 'C': 26, 'D': 23, 'E': 28},
-        28: {'A':  1, 'B': 31, 'C': 29, 'D': 16, 'E': 27},
-        29: {'A':  0, 'B': 30, 'C': 28, 'D': 17, 'E': 26},
-        30: {'A':  3, 'B': 29, 'C': 31, 'D': 18, 'E': 25},
-        31: {'A':  2, 'B': 28, 'C': 30, 'D': 19, 'E': 24},
-        }
-
-    # --------------------------------------------------------------------------
-    #
-    # BG/Q Config
-    #
-    BGQ_CORES_PER_NODE      = 16
-    BGQ_NODES_PER_BOARD     = 32 # NODE == Compute Card == Chip module
-    BGQ_BOARDS_PER_MIDPLANE = 16 # NODE BOARD == NODE CARD
-    BGQ_MIDPLANES_PER_RACK  = 2
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Default mapping = "ABCDE(T)"
-    #
-    # http://www.redbooks.ibm.com/redbooks/SG247948/wwhelp/wwhimpl/js/html/wwhelp.htm
-    #
-    BGQ_MAPPING = "ABCDE"
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Board labels (Rack, Midplane, Node)
-    #
-    BGQ_BOARD_LABELS = ['R', 'M', 'N']
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Dimensions of a (sub-)block
-    #
-    BGQ_DIMENSION_LABELS = ['A', 'B', 'C', 'D', 'E']
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Supported sub-block sizes (number of nodes).
-    # This influences the effectiveness of mixed-size allocations
-    # (and might even be a hard requirement from a topology standpoint).
-    #
-    # TODO: Do we actually need to restrict our sub-block sizes to this set?
-    #
-    BGQ_SUPPORTED_SUB_BLOCK_SIZES = [1, 2, 4, 8, 16, 32, 64, 128, 256, 512]
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Mapping of starting corners.
-    #
-    # "board" -> "node"
-    #
-    # Ordering: ['E', 'D', 'DE', etc.]
-    #
-    # TODO: Is this independent of the mapping?
-    #
-    BGQ_BLOCK_STARTING_CORNERS = {
-        0:  0,
-        4: 29,
-        8:  4,
-        12: 25
-    }
-
-
-    # --------------------------------------------------------------------------
-    #
-    # BG/Q Topology of Boards within a Midplane
-    #
-    BGQ_MIDPLANE_TOPO = {
-        0: {'A':  4, 'B':  8, 'C':  1, 'D':  2},
-        1: {'A':  5, 'B':  9, 'C':  0, 'D':  3},
-        2: {'A':  6, 'B': 10, 'C':  3, 'D':  0},
-        3: {'A':  7, 'B': 11, 'C':  2, 'D':  1},
-        4: {'A':  0, 'B': 12, 'C':  5, 'D':  6},
-        5: {'A':  1, 'B': 13, 'C':  4, 'D':  7},
-        6: {'A':  2, 'B': 14, 'C':  7, 'D':  4},
-        7: {'A':  3, 'B': 15, 'C':  6, 'D':  5},
-        8: {'A': 12, 'B':  0, 'C':  9, 'D': 10},
-        9: {'A': 13, 'B':  1, 'C':  8, 'D': 11},
-        10: {'A': 14, 'B':  2, 'C': 11, 'D':  8},
-        11: {'A': 15, 'B':  3, 'C': 10, 'D':  9},
-        12: {'A':  8, 'B':  4, 'C': 13, 'D': 14},
-        13: {'A':  9, 'B':  5, 'C': 12, 'D': 15},
-        14: {'A': 10, 'B':  6, 'C': 15, 'D': 12},
-        15: {'A': 11, 'B':  7, 'C': 14, 'D': 13},
-        }
-
-    # --------------------------------------------------------------------------
-    #
-    # Shape of whole BG/Q Midplane
-    #
-    BGQ_MIDPLANE_SHAPE = {'A': 4, 'B': 4, 'C': 4, 'D': 4, 'E': 2} # '4x4x4x4x2'
-
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        self.torus_block            = None
-        self.loadl_bg_block         = None
-        self.shape_table            = None
-        self.torus_dimension_labels = None
-
-        LRMS.__init__(self, cfg, logger)
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        loadl_node_list = None
-        loadl_cpus_per_node = None
-
-        # Determine method for determining hosts,
-        # either through hostfile or BG/Q environment.
-        loadl_hostfile = os.environ.get('LOADL_HOSTFILE')
-        self.loadl_bg_block = os.environ.get('LOADL_BG_BLOCK')
-        if loadl_hostfile is None and self.loadl_bg_block is None:
-            msg = "Neither $LOADL_HOSTFILE or $LOADL_BG_BLOCK set!"
-            self._log.error(msg)
-            raise RuntimeError(msg)
-
-        # Determine the size of the pilot allocation
-        if loadl_hostfile is not None:
-            # Non Blue Gene Load Leveler installation.
-
-            loadl_total_tasks_str = os.environ.get('LOADL_TOTAL_TASKS')
-            if loadl_total_tasks_str is None:
-                msg = "$LOADL_TOTAL_TASKS not set!"
-                self._log.error(msg)
-                raise RuntimeError(msg)
-            else:
-                loadl_total_tasks = int(loadl_total_tasks_str)
-
-            # Construct the host list
-            loadl_nodes = [line.strip() for line in open(loadl_hostfile)]
-            self._log.info("Found LOADL_HOSTFILE %s. Expanded to: %s",
-                          loadl_hostfile, loadl_nodes)
-            loadl_node_list = list(set(loadl_nodes))
-
-            # Verify that $LLOAD_TOTAL_TASKS == len($LOADL_HOSTFILE)
-            if loadl_total_tasks != len(loadl_nodes):
-                self._log.error("$LLOAD_TOTAL_TASKS(%d) != len($LOADL_HOSTFILE)(%d)",
-                               loadl_total_tasks, len(loadl_nodes))
-
-            # Determine the number of cpus per node.  Assume:
-            # cores_per_node = lenght(nodefile) / len(unique_nodes_in_nodefile)
-            loadl_cpus_per_node = len(loadl_nodes) / len(loadl_node_list)
-
-        elif self.loadl_bg_block is not None:
-            # Blue Gene specific.
-            loadl_bg_midplane_list_str = None
-            loadl_bg_block_size_str = None
-
-            loadl_job_name = os.environ.get('LOADL_JOB_NAME')
-            if loadl_job_name is None:
-                msg = "$LOADL_JOB_NAME not set!"
-                self._log.error(msg)
-                raise RuntimeError(msg)
-
-            # Get the board list and block shape from 'llq -l' output
-            output = subprocess.check_output(["llq", "-l", loadl_job_name])
-            loadl_bg_board_list_str = None
-            loadl_bg_block_shape_str = None
-            for line in output.splitlines():
-                # Detect BG board list
-                if "BG Node Board List: " in line:
-                    loadl_bg_board_list_str = line.split(':')[1].strip()
-                elif "BG Midplane List: " in line:
-                    loadl_bg_midplane_list_str = line.split(':')[1].strip()
-                elif "BG Shape Allocated: " in line:
-                    loadl_bg_block_shape_str = line.split(':')[1].strip()
-                elif "BG Size Allocated: " in line:
-                    loadl_bg_block_size_str = line.split(':')[1].strip()
-            if not loadl_bg_board_list_str:
-                msg = "No board list found in llq output!"
-                self._log.error(msg)
-                raise RuntimeError(msg)
-            self._log.debug("BG Node Board List: %s" % loadl_bg_board_list_str)
-            if not loadl_bg_midplane_list_str:
-                msg = "No midplane list found in llq output!"
-                self._log.error(msg)
-                raise RuntimeError(msg)
-            self._log.debug("BG Midplane List: %s" % loadl_bg_midplane_list_str)
-            if not loadl_bg_block_shape_str:
-                msg = "No board shape found in llq output!"
-                self._log.error(msg)
-                raise RuntimeError(msg)
-            self._log.debug("BG Shape Allocated: %s" % loadl_bg_block_shape_str)
-            if not loadl_bg_block_size_str:
-                msg = "No board size found in llq output!"
-                self._log.error(msg)
-                raise RuntimeError(msg)
-            loadl_bg_block_size = int(loadl_bg_block_size_str)
-            self._log.debug("BG Size Allocated: %d" % loadl_bg_block_size)
-
-            # Build nodes data structure to be handled by Torus Scheduler
-            try:
-                self.torus_block = self._bgq_construct_block(
-                    loadl_bg_block_shape_str, loadl_bg_board_list_str,
-                    loadl_bg_block_size, loadl_bg_midplane_list_str)
-            except Exception as e:
-                msg = "Couldn't construct block: %s" % e.message
-                self._log.error(msg)
-                raise RuntimeError(msg)
-            self._log.debug("Torus block constructed:")
-            for e in self.torus_block:
-                self._log.debug("%s %s %s %s" %
-                                (e[0], [e[1][key] for key in sorted(e[1])], e[2], e[3]))
-
-            try:
-                loadl_node_list = [entry[SchedulerTorus.TORUS_BLOCK_NAME] for entry in self.torus_block]
-            except Exception as e:
-                msg = "Couldn't construct node list."
-                self._log.error(msg)
-                raise RuntimeError(msg)
-            #self._log.debug("Node list constructed: %s" % loadl_node_list)
-
-            # Construct sub-block table
-            try:
-                self.shape_table = self._bgq_create_sub_block_shape_table(loadl_bg_block_shape_str)
-            except Exception as e:
-                msg = "Couldn't construct shape table: %s" % e.message
-                self._log.error(msg)
-                raise RuntimeError(msg)
-            self._log.debug("Shape table constructed: ")
-            for (size, dim) in [(key, self.shape_table[key]) for key in sorted(self.shape_table)]:
-                self._log.debug("%s %s" % (size, [dim[key] for key in sorted(dim)]))
-
-            # Determine the number of cpus per node
-            loadl_cpus_per_node = self.BGQ_CORES_PER_NODE
-
-            # BGQ Specific Torus labels
-            self.torus_dimension_labels = self.BGQ_DIMENSION_LABELS
-
-        self.node_list = loadl_node_list
-        self.cores_per_node = loadl_cpus_per_node
-
-        self._log.debug("Sleeping for #473 ...")
-        time.sleep(5)
-        self._log.debug("Configure done")
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Walk the block and return the node name for the given location
-    #
-    def _bgq_nodename_by_loc(self, midplanes, board, location):
-
-        self._log.debug("Starting nodebyname - midplanes:%s, board:%d" % (midplanes, board))
-
-        node = self.BGQ_BLOCK_STARTING_CORNERS[board]
-
-        # TODO: Does the order of walking matter?
-        #       It might because of the starting blocks ...
-        for dim in self.BGQ_DIMENSION_LABELS: # [::-1]:
-            max_length = location[dim]
-            self._log.debug("Within dim loop dim:%s, max_length: %d" % (dim, max_length))
-
-            cur_length = 0
-            # Loop while we are not at the final depth
-            while cur_length < max_length:
-                self._log.debug("beginning of while loop, cur_length: %d" % cur_length)
-
-                if cur_length % 2 == 0:
-                    # Stay within the board
-                    node = self.BGQ_BOARD_TOPO[node][dim]
-
-                else:
-                    # We jump to another board.
-                    self._log.debug("jumping to new board from board: %d, dim: %s)" % (board, dim))
-                    board = self.BGQ_MIDPLANE_TOPO[board][dim]
-                    self._log.debug("board is now: %d" % board)
-
-                    # If we switch boards in the B dimension,
-                    # we seem to "land" at the opposite E dimension.
-                    if dim  == 'B':
-                        node = self.BGQ_BOARD_TOPO[node]['E']
-
-                self._log.debug("node is now: %d" % node)
-
-                # Increase the length for the next iteration
-                cur_length += 1
-
-            self._log.debug("Wrapping inside dim loop dim:%s" % (dim))
-
-        # TODO: This will work for midplane expansion in one dimension only
-        midplane_idx = max(location.values()) / 4
-        rack = midplanes[midplane_idx]['R']
-        midplane = midplanes[midplane_idx]['M']
-
-        nodename = 'R%.2d-M%.1d-N%.2d-J%.2d' % (rack, midplane, board, node)
-        self._log.debug("from location %s constructed node name: %s, left at board: %d" % (self.loc2str(location), nodename, board))
-
-        return nodename
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Convert the board string as given by llq into a board structure
-    #
-    # E.g. 'R00-M1-N08,R00-M1-N09,R00-M1-N10,R00-M0-N11' =>
-    # [{'R': 0, 'M': 1, 'N': 8}, {'R': 0, 'M': 1, 'N': 9},
-    #  {'R': 0, 'M': 1, 'N': 10}, {'R': 0, 'M': 0, 'N': 11}]
-    #
-    def _bgq_str2boards(self, boards_str):
-
-        boards = boards_str.split(',')
-
-        board_dict_list = []
-
-        for board in boards:
-            elements = board.split('-')
-
-            board_dict = {}
-            for l, e in zip(self.BGQ_BOARD_LABELS, elements):
-                board_dict[l] = int(e.split(l)[1])
-
-            board_dict_list.append(board_dict)
-
-        return board_dict_list
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Convert the midplane string as given by llq into a midplane structure
-    #
-    # E.g. 'R04-M0,R04-M1' =>
-    # [{'R': 4, 'M': 0}, {'R': 4, 'M': 1}]
-    #
-    #
-    def _bgq_str2midplanes(self, midplane_str):
-
-        midplanes = midplane_str.split(',')
-
-        midplane_dict_list = []
-        for midplane in midplanes:
-            elements = midplane.split('-')
-
-            midplane_dict = {}
-            # Take the first two labels
-            for l, e in zip(self.BGQ_BOARD_LABELS[:2], elements):
-                midplane_dict[l] = int(e.split(l)[1])
-
-            midplane_dict_list.append(midplane_dict)
-
-        return midplane_dict_list
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Convert the string as given by llq into a block shape structure:
-    #
-    # E.g. '1x2x3x4x5' => {'A': 1, 'B': 2, 'C': 3, 'D': 4, 'E': 5}
-    #
-    def _bgq_str2shape(self, shape_str):
-
-        # Get the lengths of the shape
-        shape_lengths = shape_str.split('x', 4)
-
-        shape_dict = {}
-        for dim, length in zip(self.BGQ_DIMENSION_LABELS, shape_lengths):
-            shape_dict[dim] = int(length)
-
-        return shape_dict
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Multiply two shapes
-    #
-    def _multiply_shapes(self, shape1, shape2):
-
-        result = {}
-
-        for dim in self.BGQ_DIMENSION_LABELS:
-            try:
-                val1 = shape1[dim]
-            except KeyError:
-                val1 = 1
-
-            try:
-                val2 = shape2[dim]
-            except KeyError:
-                val2 = 1
-
-            result[dim] = val1 * val2
-
-        return result
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Convert location dict into a tuple string
-    # E.g. {'A': 1, 'C': 4, 'B': 1, 'E': 2, 'D': 4} => '(1,4,1,2,4)'
-    #
-    def loc2str(self, loc):
-        return str(tuple(loc[dim] for dim in self.BGQ_DIMENSION_LABELS))
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Convert a shape dict into string format
-    #
-    # E.g. {'A': 1, 'C': 4, 'B': 1, 'E': 2, 'D': 4} => '1x4x1x2x4'
-    #
-    def shape2str(self, shape):
-
-        shape_str = ''
-        for l in self.BGQ_DIMENSION_LABELS:
-
-            # Get the corresponding count
-            shape_str += str(shape[l])
-
-            # Add an 'x' behind all but the last label
-            if l in self.BGQ_DIMENSION_LABELS[:-1]:
-                shape_str += 'x'
-
-        return shape_str
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Return list of nodes that make up the block
-    #
-    # Format: [(index, location, nodename, status), (i, c, n, s), ...]
-    #
-    # TODO: This function and _bgq_nodename_by_loc should be changed so that we
-    #       only walk the torus once?
-    #
-    def _bgq_get_block(self, midplanes, board, shape):
-
-        self._log.debug("Shape: %s", shape)
-
-        nodes = []
-        index = 0
-
-        for a in range(shape['A']):
-            for b in range(shape['B']):
-                for c in range(shape['C']):
-                    for d in range(shape['D']):
-                        for e in range(shape['E']):
-                            location = {'A': a, 'B': b, 'C': c, 'D': d, 'E': e}
-                            nodename = self._bgq_nodename_by_loc(midplanes, board, location)
-                            nodes.append([index, location, nodename, FREE])
-                            index += 1
-
-        return nodes
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Use block shape and board list to construct block structure
-    #
-    # The 5 dimensions are denoted by the letters A, B, C, D, and E, T for the core (0-15).
-    # The latest dimension E is always 2, and is contained entirely within a midplane.
-    # For any compute block, compute nodes (as well midplanes for large blocks) are combined in 4 dimensions,
-    # only 4 dimensions need to be considered.
-    #
-    #  128 nodes: BG Shape Allocated: 2x2x4x4x2
-    #  256 nodes: BG Shape Allocated: 4x2x4x4x2
-    #  512 nodes: BG Shape Allocated: 1x1x1x1
-    #  1024 nodes: BG Shape Allocated: 1x1x1x2
-    #
-    def _bgq_construct_block(self, block_shape_str, boards_str,
-                            block_size, midplane_list_str):
-
-        llq_shape = self._bgq_str2shape(block_shape_str)
-
-        # TODO: Could check this, but currently _shape2num is part of the other class
-        #if self._shape2num_nodes(llq_shape) != block_size:
-        #    self._log.error("Block Size doesn't match Block Shape")
-
-        # If the block is equal to or greater than a Midplane,
-        # then there is no board list provided.
-        # But because at that size, we have only full midplanes,
-        # we can construct it.
-
-        if block_size >= 1024:
-            #raise NotImplementedError("Currently multiple midplanes are not yet supported.")
-
-            # BG Size: 1024, BG Shape: 1x1x1x2, BG Midplane List: R04-M0,R04-M1
-            midplanes = self._bgq_str2midplanes(midplane_list_str)
-
-            # Start of at the "lowest" available rack/midplane/board
-            # TODO: No other explanation than that this seems to be the convention?
-            # TODO: Can we safely assume that they are sorted?
-            #rack = midplane_dict_list[0]['R']
-            #midplane = midplane_dict_list[0]['M']
-            board = 0
-
-            # block_shape = llq_shape * BGQ_MIDPLANE_SHAPE
-            block_shape = self._multiply_shapes(self.BGQ_MIDPLANE_SHAPE, llq_shape)
-            self._log.debug("Resulting shape after multiply: %s" % block_shape)
-
-        elif block_size == 512:
-            # Full midplane
-
-            # BG Size: 1024, BG Shape: 1x1x1x2, BG Midplane List: R04-M0,R04-M1
-            midplanes = self._bgq_str2midplanes(midplane_list_str)
-
-            # Start of at the "lowest" available rack/midplane/board
-            # TODO: No other explanation than that this seems to be the convention?
-            #rack = midplane_dict_list[0]['R'] # Assume they are all equal
-            #midplane = min([entry['M'] for entry in midplane_dict_list])
-            board = 0
-
-            block_shape = self.BGQ_MIDPLANE_SHAPE
-
-        else:
-            # Within single midplane, < 512 nodes
-
-            board_dict_list = self._bgq_str2boards(boards_str)
-            self._log.debug("Board dict list:\n%s", '\n'.join([str(x) for x in board_dict_list]))
-
-            midplanes = [{'R': board_dict_list[0]['R'],
-                          'M': board_dict_list[0]['M']}]
-
-            # Start of at the "lowest" available board.
-            # TODO: No other explanation than that this seems to be the convention?
-            board = min([entry['N'] for entry in board_dict_list])
-
-            block_shape = llq_shape
-
-        # From here its all equal (assuming our walker does the walk and not just the talk!)
-        block = self._bgq_get_block(midplanes, board, block_shape)
-
-        # TODO: Check returned block:
-        #       - Length
-        #       - No duplicates
-
-        return block
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Construction of sub-block shapes based on overall block allocation.
-    #
-    # Depending on the size of the total allocated block, the maximum size
-    # of a subblock can be 512 nodes.
-    #
-    #
-    def _bgq_create_sub_block_shape_table(self, shape_str):
-
-        # Convert the shape string into dict structure
-        #
-        # For < 512 nodes: the dimensions within a midplane (AxBxCxDxE)
-        # For >= 512 nodes: the dimensions between the midplanes (AxBxCxD)
-        #
-        if len(shape_str.split('x')) == 5:
-            block_shape = self._bgq_str2shape(shape_str)
-        elif len(shape_str.split('x')) == 4:
-            block_shape = self.BGQ_MIDPLANE_SHAPE
-        else:
-            raise ValueError('Invalid shape string: %s' % shape_str)
-
-        # Dict to store the results
-        table = {}
-
-        # Create a sub-block dict with shape 1x1x1x1x1
-        sub_block_shape = {}
-        for l in self.BGQ_DIMENSION_LABELS:
-            sub_block_shape[l] = 1
-
-        # Look over all the dimensions starting at the most right
-        for dim in self.BGQ_MAPPING[::-1]:
-            while True:
-
-                # Calculate the number of nodes for the current shape
-                from operator import mul
-                num_nodes = reduce(mul, filter(lambda length: length != 0, sub_block_shape.values()))
-
-                if num_nodes in self.BGQ_SUPPORTED_SUB_BLOCK_SIZES:
-                    table[num_nodes] = copy.copy(sub_block_shape)
-                else:
-                    self._log.warning("Non supported sub-block size: %d.", num_nodes)
-
-                # Done with iterating this dimension
-                if sub_block_shape[dim] >= block_shape[dim]:
-                    break
-
-                # Increase the length in this dimension for the next iteration.
-                if sub_block_shape[dim] == 1:
-                    sub_block_shape[dim] = 2
-                elif sub_block_shape[dim] == 2:
-                    sub_block_shape[dim] = 4
-
-        return table
-
-
-
-# ==============================================================================
-#
-class ForkLRMS(LRMS):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LRMS.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        self._log.info("Using fork on localhost.")
-
-        # For the fork LRMS (ie. on localhost), we fake an infinite number of
-        # cores, so don't perform any sanity checks.
-        detected_cpus = multiprocessing.cpu_count()
-
-        if detected_cpus != self.requested_cores:
-            self._log.info("using %d instead of physically available %d cores.",
-                    self.requested_cores, detected_cpus)
-
-        # if cores_per_node is set in the agent config, we slice the number of
-        # cores into that many virtual nodes.  cpn defaults to requested_cores,
-        # to preserve the previous behavior (1 node).
-        self.cores_per_node = self._cfg.get('cores_per_node')
-        if not self.cores_per_node:
-            self.cores_per_node = self.requested_cores
-
-        requested_nodes = int(math.ceil(float(self.requested_cores) / float(self.cores_per_node)))
-        self.node_list  = list()
-        for i in range(requested_nodes):
-            self.node_list.append("localhost")
-
-        self._log.debug('configure localhost to behave as %s nodes with %s cores each.',
-                len(self.node_list), self.cores_per_node)
-
-
-
-# ==============================================================================
-#
-class YARNLRMS(LRMS):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LRMS.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        self._log.info("Using YARN on localhost.")
-
-        selected_cpus = self.requested_cores
-
-        # when we profile the agent, we fake any number of cores, so don't
-        # perform any sanity checks.  Otherwise we use at most all available
-        # cores (and informa about unused ones)
-        if 'RADICAL_PILOT_PROFILE' not in os.environ:
-
-            detected_cpus = multiprocessing.cpu_count()
-
-            if detected_cpus < selected_cpus:
-                self._log.warn("insufficient cores: using available %d instead of requested %d.",
-                        detected_cpus, selected_cpus)
-                selected_cpus = detected_cpus
-
-            elif detected_cpus > selected_cpus:
-                self._log.warn("more cores available: using requested %d instead of available %d.",
-                        selected_cpus, detected_cpus)
-
-        hdfs_conf_output =subprocess.check_output(['hdfs', 'getconf', '-nnRpcAddresses']).split('\n')[0]
-        self._log.debug('Namenode URL = {0}'.format(hdfs_conf_output))
-        self.namenode_url = hdfs_conf_output
-
-
-        self._log.debug('Namenode URL = {0}'.format(self.namenode_url))
-
-        # I will leave it for the moment because I have not found another way
-        # to take the necessary value yet.
-        yarn_conf_output = subprocess.check_output(['yarn', 'node', '-list'], stderr=subprocess.STDOUT).split('\n')
-        for line in yarn_conf_output:
-            if 'ResourceManager' in line:
-                settings = line.split('at ')[1]
-                if '/' in settings:
-                    rm_url=settings.split('/')[1]
-                    self.rm_ip=rm_url.split(':')[0]
-                    self.rm_port=rm_url.split(':')[1]
-
-                else:
-                    self.rm_ip=settings.split(':')[0]
-                    self.rm_port=settings.split(':')[1]
-
-        hostname = os.environ.get('HOSTNAME')
-
-        if hostname == None:
-            self.node_list = ['localhost']
-        else:
-            self.node_list = [hostname]
-        self.cores_per_node = selected_cpus
-
 
 
 # ==============================================================================
@@ -7152,9 +5669,9 @@ def bootstrap_3():
             # sub-agent startup.  Add the remaining LRMS information to the
             # config, for the benefit of the scheduler).
 
-            lrms = LRMS.create(name   = cfg['lrms'],
-                               cfg    = cfg,
-                               logger = log)
+            lrms = rp.agent.RM.create(name   = cfg['lrms'],
+                             cfg    = cfg,
+                             logger = log)
             cfg['lrms_info'] = lrms.lrms_info
 
 
@@ -7175,8 +5692,8 @@ def bootstrap_3():
             # FIXME: we should point the address to the node of the subagent
             #        which hosts the bridge, not the local IP.  Until this
             #        is fixed, bridges MUST run on agent_0 (which is what
-            #        LRMS.hostip() below will point to).
-            nodeip = LRMS.hostip(cfg.get('network_interface'), logger=log)
+            #        RM.hostip() below will point to).
+            nodeip = rp.agent.RM.hostip(cfg.get('network_interface'), logger=log)
             write_sub_configs(cfg, bridges, nodeip, log)
 
             # Store some runtime information into the session
