+++ src/radical/pilot/agent/rm/__init__.py
@@ -1,0 +1,10 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+# 'enum' for resource manager types
+
+from .base import *
+from .base import LRMS as RM
+
