+++ src/radical/pilot/agent/rm/ccm.py
@@ -1,0 +1,50 @@
+
+# ==============================================================================
+#
+class CCMLRMS(LRMS):
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LRMS.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+
+        self._log.info("Configured to run on system with %s.", self.name)
+
+        CCM_NODEFILE_DIR = os.path.expanduser('~/.crayccm')
+
+        ccm_nodefile_list = filter(lambda x: x.startswith('ccm_nodelist'),
+                                   os.listdir(CCM_NODEFILE_DIR))
+        if not ccm_nodefile_list:
+            raise Exception("No CCM nodefiles found in: %s." % CCM_NODEFILE_DIR)
+
+        ccm_nodefile_name = max(ccm_nodefile_list, key=lambda x:
+                              os.stat(os.path.join(CCM_NODEFILE_DIR, x)).st_mtime)
+        ccm_nodefile = os.path.join(CCM_NODEFILE_DIR, ccm_nodefile_name)
+
+        hostname = os.uname()[1]
+        if not hostname in open(ccm_nodefile).read():
+            raise RuntimeError("Using the most recent CCM nodefile (%s),"
+                               " but I (%s) am not in it!" % (ccm_nodefile, hostname))
+
+        # Parse the CCM nodefile
+        ccm_nodes = [line.strip() for line in open(ccm_nodefile)]
+        self._log.info("Found CCM nodefile: %s.", ccm_nodefile)
+
+        # Get the number of raw entries
+        ccm_nodes_length = len(ccm_nodes)
+
+        # Unique nodes
+        ccm_node_list = list(set(ccm_nodes))
+        ccm_node_list_length = len(ccm_node_list)
+
+        # Some simple arithmetic
+        self.cores_per_node = ccm_nodes_length / ccm_node_list_length
+
+        self.node_list = ccm_node_list
+
+
