+++ src/radical/pilot/agent/rm/slurm.py
@@ -1,0 +1,71 @@
+
+# ==============================================================================
+#
+class SLURMLRMS(LRMS):
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LRMS.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+
+        slurm_nodelist = os.environ.get('SLURM_NODELIST')
+        if slurm_nodelist is None:
+            msg = "$SLURM_NODELIST not set!"
+            self._log.error(msg)
+            raise RuntimeError(msg)
+
+        # Parse SLURM nodefile environment variable
+        slurm_nodes = hostlist.expand_hostlist(slurm_nodelist)
+        self._log.info("Found SLURM_NODELIST %s. Expanded to: %s", slurm_nodelist, slurm_nodes)
+
+        # $SLURM_NPROCS = Total number of cores allocated for the current job
+        slurm_nprocs_str = os.environ.get('SLURM_NPROCS')
+        if slurm_nprocs_str is None:
+            msg = "$SLURM_NPROCS not set!"
+            self._log.error(msg)
+            raise RuntimeError(msg)
+        else:
+            slurm_nprocs = int(slurm_nprocs_str)
+
+        # $SLURM_NNODES = Total number of (partial) nodes in the job's resource allocation
+        slurm_nnodes_str = os.environ.get('SLURM_NNODES')
+        if slurm_nnodes_str is None:
+            msg = "$SLURM_NNODES not set!"
+            self._log.error(msg)
+            raise RuntimeError(msg)
+        else:
+            slurm_nnodes = int(slurm_nnodes_str)
+
+        # $SLURM_CPUS_ON_NODE = Number of cores per node (physically)
+        slurm_cpus_on_node_str = os.environ.get('SLURM_CPUS_ON_NODE')
+        if slurm_cpus_on_node_str is None:
+            msg = "$SLURM_CPUS_ON_NODE not set!"
+            self._log.error(msg)
+            raise RuntimeError(msg)
+        else:
+            slurm_cpus_on_node = int(slurm_cpus_on_node_str)
+
+        # Verify that $SLURM_NPROCS <= $SLURM_NNODES * $SLURM_CPUS_ON_NODE
+        if not slurm_nprocs <= slurm_nnodes * slurm_cpus_on_node:
+            self._log.warning("$SLURM_NPROCS(%d) <= $SLURM_NNODES(%d) * $SLURM_CPUS_ON_NODE(%d)",
+                            slurm_nprocs, slurm_nnodes, slurm_cpus_on_node)
+
+        # Verify that $SLURM_NNODES == len($SLURM_NODELIST)
+        if slurm_nnodes != len(slurm_nodes):
+            self._log.error("$SLURM_NNODES(%d) != len($SLURM_NODELIST)(%d)",
+                           slurm_nnodes, len(slurm_nodes))
+
+        # Report the physical number of cores or the total number of cores
+        # in case of a single partial node allocation.
+        self.cores_per_node = min(slurm_cpus_on_node, slurm_nprocs)
+
+        self.node_list = slurm_nodes
+
+
+
